requirements.txt

flask>=3.0.0
flask-cors>=6.0.0
gunicorn>=22.0.0
openai>=1.0.0
zep-cloud==3.13.0
camel-oasis==0.2.5
camel-ai==0.2.78
PyMuPDF>=1.24.0
charset-normalizer>=3.0.0
chardet>=5.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
supabase>=2.10.0
PyJWT>=2.8.0
